Moved OpenSearchConnectionConfig to config module
